fix(B-464): cleanup session feedback schema and model
